+++ src/app/pages/choose-account/components/accounts.tsx
@@ -1,21 +1,20 @@
 import { Suspense, memo, useState, useMemo } from 'react';
 import { FiPlusCircle } from 'react-icons/fi';
 import { Virtuoso } from 'react-virtuoso';
-import { Box, BoxProps, color, FlexProps, Spinner, Stack } from '@stacks/ui';
-import { truncateMiddle } from '@stacks/ui-utils';
+import { Box, BoxProps, color, FlexProps, Stack } from '@stacks/ui';
 
-import { Caption, Text, Title } from '@app/components/typography';
+import { Text, Title } from '@app/components/typography';
 import { useAccountDisplayName } from '@app/common/hooks/account/use-account-names';
 import { useWallet } from '@app/common/hooks/use-wallet';
 import { useOnboardingState } from '@app/common/hooks/auth/use-onboarding-state';
 import { useCreateAccount } from '@app/common/hooks/account/use-create-account';
-import { AccountAvatarWithName } from '@app/components/account-avatar/account-avatar';
-import { SpaceBetween } from '@app/components/space-between';
+import { AccountAvatarWithName } from '@app/components/account/account-avatar/account-avatar';
+
 import { usePressable } from '@app/components/item-hover';
 import {
   AccountBalanceCaption,
   AccountBalanceLoading,
-} from '@app/components/account-balance-caption';
+} from '@app/components/account/account-balance-caption';
 import { slugify } from '@app/common/utils';
 import { useAccounts, useHasCreatedAccount } from '@app/store/accounts/account.hooks';
 import { useAddressBalances } from '@app/query/stacks/balance/balance.hooks';
@@ -23,6 +22,8 @@ import { useWalletType } from '@app/common/use-wallet-type';
 import { AccountWithAddress } from '@app/store/accounts/account.models';
 import { useNavigate } from 'react-router-dom';
 import { RouteUrls } from '@shared/route-urls';
+import { AccountListItemLayout } from '@app/components/account/account-list-item-layout';
+import { useStxMarketData } from '@app/query/common/market-data/market-data.hooks';
 
 const loadingProps = { color: '#A1A7B3' };
 const getLoadingProps = (loading: boolean) => (loading ? loadingProps : {});
@@ -51,64 +52,59 @@ const AccountTitle = ({ account, name, ...rest }: AccountTitleProps) => {
   );
 };
 
-interface AccountItemProps extends FlexProps {
+interface ChooseAccountItemProps extends FlexProps {
   selectedAddress?: string | null;
   isLoading: boolean;
   account: AccountWithAddress;
   onSelectAccount(index: number): void;
 }
-const AccountItem = memo((props: AccountItemProps) => {
+const ChooseAccountItem = memo((props: ChooseAccountItemProps) => {
   const { selectedAddress, account, isLoading, onSelectAccount, ...rest } = props;
   const [component, bind] = usePressable(true);
   const { decodedAuthRequest } = useOnboardingState();
   const name = useAccountDisplayName(account);
   const { data: balances, isLoading: isBalanceLoading } = useAddressBalances(account.address);
+  const fiatValue = useStxMarketData();
   const showLoadingProps = !!selectedAddress || !decodedAuthRequest;
 
   const accountSlug = useMemo(() => slugify(`Account ${account?.index + 1}`), [account?.index]);
 
   return (
-    <Box
-      height="68px"
-      data-testid={`account-${accountSlug}-${account.index}`}
-      onClick={() => onSelectAccount(account.index)}
-      {...rest}
-    >
-      <Box {...bind}>
-        <Stack spacing="base" alignSelf="center" isInline>
-          <AccountAvatarWithName name={name} flexGrow={0} publicKey={account.stxPublicKey} />
-          <SpaceBetween width="100%" alignItems="center">
-            <Stack textAlign="left" spacing="base-tight">
-              <Suspense
-                fallback={
-                  <AccountTitlePlaceholder
-                    {...getLoadingProps(showLoadingProps)}
-                    account={account}
-                  />
-                }
-              >
-                <AccountTitle
-                  name={name}
-                  {...getLoadingProps(showLoadingProps)}
-                  account={account}
-                />
-              </Suspense>
-              <Stack alignItems="center" spacing="6px" isInline>
-                <Caption fontSize={0} {...getLoadingProps(showLoadingProps)}>
-                  {truncateMiddle(account.address, 4)}
-                </Caption>
-                {isBalanceLoading ? (
-                  <AccountBalanceLoading />
-                ) : (
-                  balances && <AccountBalanceCaption availableBalance={balances.availableStx} />
-                )}
-              </Stack>
-            </Stack>
-            {isLoading && <Spinner width={4} height={4} {...loadingProps} />}
-          </SpaceBetween>
-        </Stack>
+    // Padding required on outer element to prevent jumpy list behaviours in
+    // virtualised list library
+    <Box pb="loose">
+      <AccountListItemLayout
+        account={account}
+        accountName={
+          <Suspense
+            fallback={
+              <AccountTitlePlaceholder {...getLoadingProps(showLoadingProps)} account={account} />
+            }
+          >
+            <AccountTitle name={name} {...getLoadingProps(showLoadingProps)} account={account} />
+          </Suspense>
+        }
+        avatar={<AccountAvatarWithName name={name} flexGrow={0} publicKey={account.stxPublicKey} />}
+        balanceLabel={
+          isBalanceLoading ? (
+            <AccountBalanceLoading />
+          ) : balances ? (
+            <AccountBalanceCaption
+              availableBalance={balances.availableStx}
+              marketData={fiatValue}
+            />
+          ) : (
+            <></>
+          )
+        }
+        isLoading={isLoading}
+        onSelectAccount={() => onSelectAccount(account.index)}
+        data-testid={`account-${accountSlug}-${account.index}`}
+        {...bind}
+        {...rest}
+      >
         {component}
-      </Box>
+      </AccountListItemLayout>
     </Box>
   );
 });
@@ -134,7 +130,7 @@ const AddAccountAction = memo(() => {
   );
 });
 
-export const Accounts = memo(() => {
+export const ChooseAccountsList = memo(() => {
   const { finishSignIn } = useWallet();
   const { whenWallet } = useWalletType();
   const accounts = useAccounts();
@@ -161,9 +157,8 @@ export const Accounts = memo(() => {
       <Virtuoso
         useWindowScroll
         data={accounts}
-        style={{ height: '68px' }}
         itemContent={(index, account) => (
-          <AccountItem
+          <ChooseAccountItem
             account={account}
             isLoading={whenWallet({ software: selectedAccount === index, ledger: false })}
             onSelectAccount={signIntoAccount}