+++ src/app/pages/send-tokens/components/send-tokens-confirm-drawer/send-tokens-confirm-drawer.tsx
@@ -1,5 +1,6 @@
 import { useFormikContext } from 'formik';
 import { Stack } from '@stacks/ui';
+import BigNumber from 'bignumber.js';
 
 import { useDrawers } from '@app/common/hooks/use-drawers';
 import { BaseDrawer, BaseDrawerProps } from '@app/components/drawer/base-drawer';
@@ -10,11 +11,20 @@ import { TransactionFee } from '@app/components/fee-row/components/transaction-f
 import { useSendFormUnsignedTxPreviewState } from '@app/store/transactions/transaction.hooks';
 
 import { SendTokensConfirmActions } from './send-tokens-confirm-actions';
-import { useEffect } from 'react';
+import { useEffect, useMemo } from 'react';
 import { SendTokensConfirmDetails } from './send-tokens-confirm-details';
 import { useAnalytics } from '@app/common/hooks/analytics/use-analytics';
 import { SendFormValues } from '@app/common/transactions/transaction-utils';
 import { StacksTransaction } from '@stacks/transactions';
+import { useConvertStxToFiatAmount } from '@app/common/hooks/use-convert-to-fiat-amount';
+import { createMoney } from '@shared/models/money.model';
+import { microStxToStx } from '@app/common/stacks-utils';
+
+function getFeeWithDefaultOfZero(tx?: StacksTransaction) {
+  if (!tx) return createMoney(0, 'STX');
+  const fee = microStxToStx(Number(tx.auth.spendingCondition.fee));
+  return createMoney(new BigNumber(fee), 'STX');
+}
 
 interface SendTokensSoftwareConfirmDrawerProps extends BaseDrawerProps {
   onUserSelectBroadcastTransaction(tx: StacksTransaction | undefined): void;
@@ -22,16 +32,23 @@ interface SendTokensSoftwareConfirmDrawerProps extends BaseDrawerProps {
 export function SendTokensSoftwareConfirmDrawer(props: SendTokensSoftwareConfirmDrawerProps) {
   const { isShowing, onClose, onUserSelectBroadcastTransaction } = props;
   const { values } = useFormikContext<SendFormValues>();
-  const transaction = useSendFormUnsignedTxPreviewState(values.assetId, values);
+  const unsignedTransaction = useSendFormUnsignedTxPreviewState(values.assetId, values);
   const analytics = useAnalytics();
   const { showEditNonce } = useDrawers();
+
+  const convertStxToUsd = useConvertStxToFiatAmount();
+
+  const feeInUsd = useMemo(
+    () => convertStxToUsd(getFeeWithDefaultOfZero(unsignedTransaction)),
+    [convertStxToUsd, unsignedTransaction]
+  );
 
   useEffect(() => {
     if (!isShowing) return;
     void analytics.track('view_transaction_signing');
   }, [isShowing, analytics]);
 
-  if (!isShowing || !transaction || !values) return null;
+  if (!isShowing || !unsignedTransaction || !values) return null;
 
   return (
     <BaseDrawer
@@ -40,22 +57,28 @@ export function SendTokensSoftwareConfirmDrawer(props: SendTokensSoftwareConfirm
       onClose={onClose}
       pauseOnClickOutside={showEditNonce}
     >
-      <Stack pb="extra-loose" px="loose" spacing="loose">
+      <Stack pb="extra-loose" px="loose" spacing="base">
         <SendTokensConfirmDetails
           amount={values.amount}
           assetId={values.assetId}
           recipient={values.recipient}
-          nonce={Number(transaction?.auth.spendingCondition?.nonce)}
         />
         <SpaceBetween>
           <Caption>Fees</Caption>
           <Caption>
-            <TransactionFee fee={values.fee} />
+            <TransactionFee
+              fee={getFeeWithDefaultOfZero(unsignedTransaction).amount.toString()}
+              usdAmount={feeInUsd}
+            />
           </Caption>
         </SpaceBetween>
+        <SpaceBetween>
+          <Caption>Nonce</Caption>
+          <Caption>{Number(unsignedTransaction.auth.spendingCondition.nonce)}</Caption>
+        </SpaceBetween>
         <SendTokensConfirmActions
-          transaction={transaction}
-          onUserConfirmBroadcast={() => onUserSelectBroadcastTransaction(transaction)}
+          transaction={unsignedTransaction}
+          onUserConfirmBroadcast={() => onUserSelectBroadcastTransaction(unsignedTransaction)}
         />
       </Stack>
     </BaseDrawer>