+++ src/app/pages/send-tokens/send-tokens.tsx
@@ -52,7 +52,7 @@ function SendTokensFormBase() {
   const { whenWallet } = useWalletType();
   const ledgerNavigate = useLedgerNavigate();
 
-  useRouteHeader(<Header title="Send" onClose={() => navigate(RouteUrls.Home)} />);
+  useRouteHeader(<Header hideActions title="Send" onClose={() => navigate(RouteUrls.Home)} />);
 
   useEffect(() => {
     if (showNetworks) navigate(RouteUrls.Home);
@@ -134,7 +134,7 @@ function SendTokensFormBase() {
               <SendFormInner
                 assetError={assetError}
                 feeEstimations={feeEstimations.estimates}
-                onAssetIdSelected={(assetId: string) => setSelectedAssetId(assetId)}
+                onAssetIdSelected={assetId => setSelectedAssetId(assetId)}
                 nonce={nonce}
               />
             </Suspense>